pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lyriclabel"
version = "0.1.0"
description = "Fetch and embed song metadata into MP3 files using the Last.fm API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",
    "mutagen>=1.46.0",
    "python-dotenv>=1.1.0",
]

[project.scripts]
lyriclabel = "main:main"

[tool.uv]
dev-dependencies = []
